requirements.txt
langchain==0.1.17
langchain-chroma==0.1.1
langchain-community==0.0.36
langchain-core==0.1.50
langchain-experimental==0.0.57
langchain-openai==0.1.6
langchain-pinecone==0.1.0
langchain-text-splitters==0.0.1
langcodes==3.4.0
langdetect==1.0.9
langsmith==0.1.53
language_data==1.2.0
Languages==1.0.0
pandas==2.2.2
pandas-profiling==3.2.0
psycopg2-binary
python-dotenv
streamlit
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-pills==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
